refactor: inline OutputConfigurationExtension directly into Infrastructure
Remove the OutputConfigurationExtension<T> wrapper class and store the
configuration list directly in Infrastructure<TOutputConfiguration, TMetadata>.
The extension was a pure data holder with no lifecycle hooks, making the
indirection unnecessary.
